requirements.txt
requests>=2.31.0
pandas>=2.1.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pytz>=2023.3